moved ica query
