+++ x/stakeibc/keeper/grpc_query.go
@@ -11,6 +11,8 @@ import (
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 
+	icatypes "github.com/cosmos/ibc-go/v7/modules/apps/27-interchain-accounts/types"
+
 	"github.com/Stride-Labs/stride/v18/x/stakeibc/types"
 )
 
@@ -69,6 +71,23 @@ func (k Keeper) Validators(c context.Context, req *types.QueryGetValidatorsReque
 	}
 
 	return &types.QueryGetValidatorsResponse{Validators: hostZone.Validators}, nil
+}
+
+// InterchainAccountFromAddress implements the Query/InterchainAccountFromAddress gRPC method
+func (k Keeper) InterchainAccountFromAddress(goCtx context.Context, req *types.QueryInterchainAccountFromAddressRequest) (*types.QueryInterchainAccountFromAddressResponse, error) {
+	ctx := sdk.UnwrapSDKContext(goCtx)
+
+	portID, err := icatypes.NewControllerPortID(req.Owner)
+	if err != nil {
+		return nil, status.Errorf(codes.InvalidArgument, "could not find account: %s", err)
+	}
+
+	addr, found := k.ICAControllerKeeper.GetInterchainAccountAddress(ctx, req.ConnectionId, portID)
+	if !found {
+		return nil, status.Errorf(codes.NotFound, "no account found for portID %s", portID)
+	}
+
+	return types.NewQueryInterchainAccountResponse(addr), nil
 }
 
 func (k Keeper) AllTradeRoutes(c context.Context, req *types.QueryAllTradeRoutes) (*types.QueryAllTradeRoutesResponse, error) {
